# src/main/resources/assets/api-docs.yaml
openapi: 3.0.3
info:
  title: Consent
  description: A set of web-services containing methods for ingesting and retrieving expressions representing data use restrictions found within consent forms.
  version: 1.0.1
  license:
    name: BSD
    url: http://opensource.org/licenses/BSD-3-Clause
  termsOfService: https://github.com/DataBiosphere/consent
servers:
  - url: /
security:
  - googleoauth:
      - email
      - profile
paths:
...
  /api/dar/v2:
    post:
      summary: Create Data Access Request, version 2
      description: Creates a Data Access Request
      parameters:
        - name: dar
          in: body
          description: The fields that represent a DAR in json format.
          required: true
          schema:
            $ref: '#/components/schemas/DataAccessRequest'
      tags:
        - Data Access Request
      responses:
        200:
          description: Returns the created Data Access Request.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DataAccessRequest'
        500:
          description: Internal Server Error.
  /api/dar/v2/{referenceId}:
    get:
      summary: Get a Data Access Request by Reference Id
      description: Returns the created Data Access Request.
      parameters:
        - name: referenceId
          in: path
          description: The referenceId of the Data Access Request
          required: true
          schema:
            type: string
      tags:
        - Data Access Request
      responses:
        200:
          description: Returns a Data Access Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataAccessRequest'
        400:
          description: The provided Reference Id is not a valid DAR identifier
        404:
          description: No DAR can be found with the provided identifier
        500:
          description: Server Error
    put:
      summary: Update a Data Access Request by Reference Id
      description: Returns the updated Data Access Request.
      parameters:
        - name: referenceId
          in: path
          description: The referenceId of the Data Access Request
          required: true
          schema:
            type: string
        - name: dar
          in: body
          description: The fields that represent a DAR in json format.
          required: true
          schema:
            $ref: '#/components/schemas/DataAccessRequest'
      tags:
        - Data Access Request
      responses:
        200:
          description: Returns a Data Access Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataAccessRequest'
        403:
          description: Updates are restricted to the create user.
        500:
          description: Internal Server Error.
  /api/dar/v2/draft:
    post:
      summary: Create Draft Data Access Request
      description: Create Draft Data Access Request
      parameters:
        - name: dar
          in: body
          description: The fields that represent a Data Access Request, json format.
          required: true
          schema:
            $ref: '#/components/schemas/DataAccessRequest'
      tags:
        - Data Access Request
      responses:
        201:
          description: Returns the created draft Data Access Request.
        500:
          description: Internal Server Error.
  /api/dar/v2/draft/{referenceId}:
    put:
      summary: Update Draft Data Access Request
      description: Update Draft Data Access Request
      parameters:
        - name: referenceId
          in: path
          description: The referenceId of the Data Access Request
          required: true
          schema:
            type: string
        - name: dar
          in: body
          description: The fields that represent a Data Access Request, json format.
          required: true
          schema:
            $ref: '#/components/schemas/DataAccessRequest'
      tags:
        - Data Access Request
      responses:
        201:
          description: Returns the upated draft Data Access Request.
        403:
          description: Updates are restricted to the create user.
        500:
          description: Internal Server Error.
  /api/dar/v2/{referenceId}/irbDocument:
    get:
      summary: Retrieve DAR IRB Document
      description: Retrieve the IRB Document for a Data Access Request
      parameters:
        - name: referenceId
          in: path
          description: The referenceId of the Data Access Request
          required: true
          schema:
            type: string
      tags:
        - Data Access Request
      responses:
        200:
          description: Returns the document
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Store DAR IRB Document
      description: |
        Store an IRB Document for a Data Access Request.
        Previous document is deleted.
      parameters:
        - name: referenceId
          in: path
          description: The referenceId of the Data Access Request
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      tags:
        - Data Access Request
      responses:
        200:
          description: Stores the document and returns the updated Data Access Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataAccessRequest'
  /api/dar/v2/{referenceId}/collaborationDocument:
    get:
      summary: Retrieve DAR Collaboration Document
      description: Retrieve the Collaboration Document for a Data Access Request
      parameters:
        - name: referenceId
          in: path
          description: The referenceId of the Data Access Request
          required: true
          schema:
            type: string
      tags:
        - Data Access Request
      responses:
        200:
          description: Returns the document
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Store DAR Collaboration Document
      description: |
        Store a Collaboration Document for a Data Access Request.
        Previous document is deleted.
      parameters:
        - name: referenceId
          in: path
          description: The referenceId of the Data Access Request
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      tags:
        - Data Access Request
      responses:
        200:
          description: Stores the document and returns the updated Data Access Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataAccessRequest'
...
  /api/electionReview/access/{electionId}:
    get:
      summary: describeElectionById
      description: Returns the ElectionReview from an Election identified by ElectionId. This endpoint can be used only for DataAccess elections.
      parameters:
        - name: electionId
          in: path
          description: Election Id
          required: true
          schema:
            type: integer
        - name: isFinalAccess
          in: query
          description: Defines if the reviewVote array of the retrieved ElectionReview will contain or not the isFinalAccess = TRUE Vote.
          required: true
          schema:
            type: integer
      tags:
        - ElectionReview
      responses:
        200:
          description: ElectionReview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElectionReview'
...
components:
  securitySchemes:
    googleoauth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            email: email authorization
            profile: profile authorization
  schemas:
...
    Election:
      type: object
      properties:
        electionId:
          type: integer
          format: int32
          description: Describes the id of the election.
        electionType:
          type: string
          description: Describes the type of the election.
        finalVote:
          type: boolean
          description: Describes the chairperson final vote, if the election is closed.
        status:
          type: string
          description: Describes the status of the election. Open, Closed or Cancelled.
        createDate:
          type: string
          format: date
          description: Describes the date the election was created.
        lastUpdate:
          type: string
          format: date
          description: Describes the date the election was last modified.
        finalVoteDate:
          type: string
          format: date
          description: Describes the date the chaiperson made the final vote.
        referenceId:
          type: string
          description: Describes the id of the consent/data access request the election is related to.
        finalRationale:
          type: string
          description: If the final vote is NO, the chairperson can specify why in this field.
        finalAccessVote:
          type: boolean
          description: For Data Access Request only. Final vote that determines if the request is approved or denied.
        useRestriction:
          type: string
          description: The use restriction related to the election.
        translatedUseRestriction:
          type: string
          description: The translated use restriction related to the election.
        dataSetId:
          type: integer
          description: The dataSetId related to the election.
        displayId:
          type: string
          description: The ID for display.
        dulName:
          type: string
          description: Contains the correspondant Dul name for the given election.
        version:
          type: integer
          format: int32
          description: Indicates the version number of the given election.
        archived:
          type: boolean
          description: Determines if the election is archived.
...
    Collaborator:
      type: object
      description: |
        A Collaborator on a Data Access Request. Collaborators are either internal lab staff or
        internal collaborators.
      properties:
        approverStatus:
          type: boolean
          description: |
            Collaborator downloader/approver status. This is relevant for internal lab staff that
            will have a status as "Designated Downloader/Approver". This indication is limited to
            individuals who the PI designates to download and/or share the requested data with other
            internal lab staff, i.e. staff members and trainees under the direct supervision of the PI
        email:
          type: string
          description: Collaborator email
        eraCommonsId:
          type: string
          description: Collaborator era commons id
        name:
          type: string
          description: Collaborator name
        title:
          type: string
          description: Collaborator title
        uuid:
          type: string
          description: Optional uuid for this collaborator
    DataAccessRequest:
      type: object
      description: Data Access Request
      properties:
        referenceId:
          type: string
          description: The reference id for this DAR. Not modifiable once created.
        userId:
          type: integer
          description: Describes the id of the User. Not modifiable once created.
        createDate:
          type: number
          description: Creation Date
        sortDate:
          type: number
          description: Sort Date
        submissionDate:
          type: number
          description: Submission Date
        updateDate:
          type: number
          description: Update Date
        draft:
          type: boolean
          description: Draft status of this DAR
        darCode:
          type: string
          description: Unique DAR Code
        partialDarCode:
          type: string
          description: Unique Partial DAR Code
        projectTitle:
          type: string
          description: Project Title
        datasetIds:
          type: array
          items:
            type: string
          description: Array of Dataset Ids
        rus:
          type: string
          description: Describes the Research Use Statement
        nonTechRus:
          type: string
          description: Describes the Non Technical Summary
        diseases:
          type: boolean
          description: Defines if the purpose of the study is disease-related
        methods:
          type: boolean
          description: Defines if the purpose of the study is for Methods development
        controls:
          type: boolean
          description: Defines if the purpose of the study is to increase the number of controls available for a comparison group
        population:
          type: boolean
          description: Defines if the purpose of the study is to understand variation in the general population
        other:
          type: boolean
          description: Defines if the purpose of the study is other than the above
        otherText:
          type: string
          description: Describes above field
        ontologies:
          type: array
          description: Array of ontology terms
          items:
            type: string
        forProfit:
          type: boolean
          description: Describes if data will be used for commercial purposes
        oneGender:
          type: boolean
          description: Describes if study is gender specific
        gender:
          type: string
          description: F/M if oneGender is specified
        pediatric:
          type: boolean
          description: Describes if study is restricted to a pediatric population
        illegalBehavior:
          type: boolean
          description: If TRUE, triggers MANUAL REVIEW.
        addiction:
          type: boolean
          description: If TRUE, triggers MANUAL REVIEW
        sexualDiseases:
          type: boolean
          description: If TRUE, triggers MANUAL REVIEW
        stigmatizedDiseases:
          type: boolean
          description: If TRUE, triggers MANUAL REVIEW
        vulnerablePopulation:
          type: boolean
          description: If TRUE, triggers MANUAL REVIEW
        populationMigration:
          type: boolean
          description: If TRUE, triggers MANUAL REVIEW
        psychiatricTraits:
          type: boolean
          description: If TRUE, triggers MANUAL REVIEW
        notHealth:
          type: boolean
          description: If TRUE, triggers MANUAL REVIEW
        hmb:
          type: boolean
          description: HMB Research
        poa:
          type: boolean
          description: Populations Origins Ancestry Research
        anvilUse:
          type: boolean
          description: Anvil Use
        cloudUse:
          type: boolean
          description: Cloud Use
        cloudProvider:
          type: string
          description: Cloud Use Provider
        cloudProviderType:
          type: string
          description: Cloud Use Provider Type
        geneticStudiesOnly:
          type: boolean
          description: Genetic Studies Only
        irb:
          type: boolean
          description: IRB Approval Required
        irbDocumentLocation:
          type: string
          description: IRB Document Location
        irbProtocolExpiration:
          type: string
          description: IRB Protocol Expiration
        publication:
          type: boolean
          description: Publication Required
        collaboration:
          type: boolean
          description: Collaboration Required
        collaborationLetterLocation:
          type: string
          description: Collaboration Letter Location
        forensicActivities:
          type: boolean
          description: Study involves Forensic Activities
        sharingDistribution:
          type: boolean
          description: Study requires sharing distribution
        externalCollaborators:
          type: array
          description: |
            External Collaborators are not employees of the Requesting PI's institution and/or do not
            work at the same location as the PI, and consequently must be independently approved to
            access controlled-access data subject to the GDS Policy. External Collaborators must be
            at the PI or equivalent level and are not required to have a  Library Card in order to
            access data, although it is encouraged. Note, External Collaborators must submit an
            independant DAR approved by their Signing Official to collaborate on this project.
            External Collaborators will be able to add their Lab Staff, as needed, via their
            independent DAR. Approval of this DAR does not indicate approval of the External
            Collaborators listed.
          items:
            $ref: '#/components/schemas/Collaborator'
        internalCollaborators:
          type: array
          description: |
            Internal Collaborators are defined as individuals not under the direct supervision of the
            PO's research project involving controlled-access data subject to the NIH GDS Policy.
            They are employees of the requesting PI's institution and work at the same location/campus
            as the PI. Internal Collaborators must be at the PI or equivalent level and are required
            to have a Library Card in order to access data through this request. Internal
            Collaborators will have Data Downloader/Approval status so that they may add their own
            relevant Internal Lab Staff. Internal Collaborators will not be required to submit an
            independent DAR to collaborate on this project.
          items:
            $ref: '#/components/schemas/Collaborator'
        labCollaborators:
          type: array
          description: |
            Lab Collaborators are defined as users of data from this Data Access Request,
            including any data that are downloaded or utilized in the cloud.
          items:
            $ref: '#/components/schemas/Collaborator'
...
    ElectionReview:
      type: object
      properties:
        reviewVote:
          type: array
          items:
            $ref: '#/components/schemas/ElectionReviewVote'
          description: Votes related to the Election.
        election:
          $ref: '#/components/schemas/Election'
        consent:
          $ref: '#/components/schemas/Consent'
        voteAgreement:
          $ref: '#/components/schemas/Vote'
        finalVote:
          $ref: '#/components/schemas/Vote'
        rpElectionId:
          type: integer
          description: The research purpose ID.
    ElectionReviewVote:
      type: object
      properties:
        Vote:
          $ref: '#/components/schemas/Vote'
        displayName:
          type: string
          description: Name of the voter.
        email:
          type: string
          description: Email of the voter.
...
    ErrorResponse:
      properties:
        message:
          type: string
          description: The error message
        code:
          type: integer
          description: The response status code
...